fix(security): satisfy decoder and task gates
issue #2082

Task: SECURITY-002

Lane: L2
